# docs/OMNICUS_CRM_OUTBOUND_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Omnicus API for Cyber Pulse CRM
  version: 4.0.0
  description: Project-scoped CRM to Omnicus Telegram and WhatsApp messaging, templates, durable mutations, media and channel-aware capability contract.
security:
  - serviceBearer: []
paths:
  /integrations/v1/crm/media:
    post:
      operationId: uploadCrmOutboundMedia
      summary: Validate and durably store media before sending it
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required:
                - crmProjectId
                - omnicusProjectId
                - kind
                - file
              properties:
                crmProjectId:
                  type: string
                  minLength: 1
                  maxLength: 128
                omnicusProjectId:
                  type: string
                  minLength: 1
                  maxLength: 128
                channel:
                  type: string
                  enum: [telegram, whatsapp]
                  default: telegram
                  description: Contract 4 discriminator. Omission remains Telegram-compatible for v3 clients.
                kind:
                  $ref: '#/components/schemas/MediaKind'
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Validated media asset available for a subsequent outbound request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedMedia'
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
  /integrations/v1/crm/messages/outbound:
    post:
      operationId: queueCrmOutboundMessage
      summary: Durably queue a channel-aware Telegram or WhatsApp message from CRM
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutboundMessage'
      responses:
        '200':
          description: The intent is durably queued, not necessarily sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuedOperation'
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
  /integrations/v1/crm/capabilities:
    get:
      operationId: getCrmChannelCapabilities
      summary: Read connection-scoped Telegram or WhatsApp capabilities and limits
      description: >-
        messageEffects.limits.availableEffects is intentionally empty because Telegram Bot API
        10.2 has no bot method for discovering effect IDs. editMessage.limits defines editable,
        immutable, preserved, and omission semantics. userReactionEvents is enabled after live
        verification of add/change/remove, duplicate delivery, reaction-before-source, routing
        isolation and the post-fix transactional worker result.
      parameters:
        - $ref: '#/components/parameters/CrmProjectId'
        - $ref: '#/components/parameters/OmnicusProjectId'
        - name: connectionId
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: channel
          in: query
          required: false
          schema: { type: string, enum: [telegram, whatsapp], default: telegram }
        - name: omnicusContactId
          in: query
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: channelIdentityId
          in: query
          schema: { type: string, minLength: 1, maxLength: 128 }
      responses:
        '200':
          description: Machine-readable capabilities; absent capabilities must never be inferred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
  /integrations/v1/crm/message-templates:
    get:
      operationId: listCrmWhatsAppMessageTemplates
      summary: List normalized WhatsApp templates for one isolated conversation route
      parameters:
        - $ref: '#/components/parameters/CrmProjectId'
        - $ref: '#/components/parameters/OmnicusProjectId'
        - name: connectionId
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: omnicusContactId
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: channelIdentityId
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: channel
          in: query
          required: true
          schema: { const: whatsapp }
        - name: status
          in: query
          schema:
            type: string
            enum: [APPROVED, PENDING, REJECTED, PAUSED, DISABLED, UNKNOWN]
      responses:
        '200':
          description: Safe normalized templates without provider payload or credentials.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/WhatsAppMessageTemplate' }
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
  /integrations/v1/crm/messages/{messageId}/read:
    put:
      operationId: markCrmWhatsAppMessageRead
      summary: Durably queue a WhatsApp read marker with stable idempotency
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WhatsAppScope' }
      responses:
        '200':
          description: Durable operation is queued or idempotently replayed; UNKNOWN requires reconciliation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueuedOperation' }
        '400': { $ref: '#/components/responses/SafeError' }
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
        '409': { $ref: '#/components/responses/SafeError' }
  /integrations/v1/crm/chat-actions:
    post:
      operationId: sendCrmTelegramChatAction
      summary: Send an ephemeral five-second Telegram chat action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TelegramScope'
                - type: object
                  required: [action]
                  properties:
                    action:
                      type: string
                      enum:
                        [
                          TYPING,
                          RECORD_VOICE,
                          UPLOAD_PHOTO,
                          UPLOAD_VIDEO,
                          UPLOAD_DOCUMENT,
                          RECORD_VIDEO_NOTE,
                        ]
      responses:
        '200': { $ref: '#/components/responses/EphemeralAccepted' }
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
  /integrations/v1/crm/drafts:
    post:
      operationId: updateCrmTelegramDraft
      summary: Update a private-chat 30-second Telegram draft preview
      description: >-
        The preview is ephemeral and never SENT. Empty text/rich content is ignored. Rich media
        can reuse only a same-connection Telegram media ID because sendRichMessageDraft rejects
        direct uploads. Final content must use messages/outbound; previews expire within 30 seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TelegramScope'
                - type: object
                  required: [draftId]
                  properties:
                    draftId: { type: integer, minimum: 1 }
                    text: { type: string, maxLength: 4096 }
                    entities:
                      type: array
                      maxItems: 100
                      items: { $ref: '#/components/schemas/MessageEntity' }
                    richMessage: { $ref: '#/components/schemas/RichMessage' }
      responses:
        '200': { $ref: '#/components/responses/DraftUpdate' }
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
  /integrations/v1/crm/messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateCrmTelegramMessage
      description: >-
        Only text, caption, entities, inlineKeyboard and linkPreviewOptions are editable.
        protectContent, messageEffectId, reply/quote metadata are immutable and preserved locally.
        Omitting inlineKeyboard preserves it; replacing text/caption without entities clears prior
        formatting; omitted linkPreviewOptions uses Telegram edit defaults.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TelegramScope'
                - type: object
                  properties:
                    text: { type: string, minLength: 1, maxLength: 4096 }
                    caption: { type: string, maxLength: 1024 }
                    entities:
                      type: array
                      maxItems: 100
                      items: { $ref: '#/components/schemas/MessageEntity' }
                    inlineKeyboard: { $ref: '#/components/schemas/InlineKeyboard' }
                    linkPreviewOptions: { $ref: '#/components/schemas/LinkPreviewOptions' }
      responses:
        '200': { $ref: '#/components/responses/QueuedOperationResponse' }
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
        '409': { $ref: '#/components/responses/SafeError' }
    delete:
      operationId: deleteCrmTelegramMessage
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramScope' }
      responses:
        '200': { $ref: '#/components/responses/QueuedOperationResponse' }
        '401': { $ref: '#/components/responses/SafeError' }
        '404': { $ref: '#/components/responses/SafeError' }
        '409': { $ref: '#/components/responses/SafeError' }
...
  /integrations/v1/crm/messages/{messageId}/pin:
    parameters:
      - name: messageId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CorrelationId'
    put:
      operationId: pinCrmTelegramMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TelegramScope'
                - type: object
                  properties:
                    disableNotification: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/QueuedOperationResponse' }
        '409': { $ref: '#/components/responses/SafeError' }
    delete:
      operationId: unpinCrmTelegramMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramScope' }
      responses:
        '200': { $ref: '#/components/responses/QueuedOperationResponse' }
        '409': { $ref: '#/components/responses/SafeError' }
...
  /integrations/v1/crm/messages/media-group:
    post:
      operationId: queueCrmTelegramMediaGroup
      description: Queues one durable aggregate and one Telegram sendMediaGroup call.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaGroup' }
      responses:
        '200':
          description: Aggregate queued; not necessarily sent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaGroupQueuedOperation' }
        '409': { $ref: '#/components/responses/SafeError' }
...
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CrmProjectId:
      name: crmProjectId
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...
  schemas:
...
    MediaKind:
      type: string
      description: WhatsApp supports PHOTO, DOCUMENT, VIDEO, AUDIO, VOICE and static STICKER only.
      enum:
        - PHOTO
        - DOCUMENT
        - VIDEO
        - AUDIO
        - VOICE
        - VIDEO_NOTE
        - ANIMATION
        - STICKER
...
    UploadedMedia:
      type: object
      additionalProperties: false
      required:
        - mediaAssetId
        - kind
        - status
        - sizeBytes
        - validationChannel
      properties:
        mediaAssetId:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/MediaKind'
        status:
          type: string
          enum:
            - AVAILABLE
        sizeBytes:
          type: string
          pattern: '^[0-9]+$'
        validationChannel:
          type: string
          enum: [telegram, whatsapp]
          description: Provider rules used to validate the stored bytes; raw provider metadata is never returned.
...
  responses:
...
    SafeError:
      description: Safe structured error without provider payload or credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'